+++ src/rt/model.h
@@ -57,7 +57,8 @@ void watch_free(rt_model_t *m, rt_watch_t *w);
 
 void model_set_phase_cb(rt_model_t *m, model_phase_t phase, rt_event_fn_t fn,
                         void *user);
-rt_watch_t *model_set_event_cb(rt_model_t *m, rt_signal_t *s, rt_watch_t *w);
+rt_watch_t *model_set_event_cb(rt_model_t *m, rt_signal_t *s, int offset,
+                               int count, rt_watch_t *w);
 void model_set_timeout_cb(rt_model_t *m, uint64_t when, rt_event_fn_t fn,
                           void *user);