+++ test/vhpi/vhpi_test.h
@@ -71,5 +71,6 @@ void issue1192_startup(void);
 void issue1213_startup(void);
 void issue1233_startup(void);
 void issue1240_startup(void);
+void issue1428_startup(void);
 
 #endif  // _VHPI_TEST_H