+++ NEWS.md
@@ -3,6 +3,7 @@
   VHPI extensions.
 - Added a `vhpiRandomSeedP` property for `vhpiToolK` which returns the
   random seed for the simulation.
+- VHPI value change callbacks for indexed names now work correctly (#1428).
 
 ## Version 1.19.2 - 2026-02-22
 - Fixed a crash when dumping a waveform `--dump-arrays` and the design